+++ src/settings/SettingsTab.ts
@@ -4,20 +4,25 @@ import { DISPLAY_NAMES } from "./languageDisplayName";
 import makeCppSettings from "./per-lang/makeCppSettings";
 import makeCSettings from "./per-lang/makeCSettings.js";
 import makeCsSettings from "./per-lang/makeCsSettings";
+import makeFSharpSettings from "./per-lang/makeFSharpSettings";
 import makeGoSettings from "./per-lang/makeGoSettings";
 import makeGroovySettings from "./per-lang/makeGroovySettings";
 import makeHaskellSettings from "./per-lang/makeHaskellSettings";
 import makeJavaSettings from "./per-lang/makeJavaSettings";
 import makeJsSettings from "./per-lang/makeJsSettings";
 import makeKotlinSettings from "./per-lang/makeKotlinSettings";
+import makeLeanSettings from "./per-lang/makeLeanSettings";
 import makeLuaSettings from "./per-lang/makeLuaSettings";
+import makeDartSettings from "./per-lang/makeDartSettings";
 import makeMathematicaSettings from "./per-lang/makeMathematicaSettings";
 import makePowershellSettings from "./per-lang/makePowershellSettings";
 import makePrologSettings from "./per-lang/makePrologSettings";
 import makePythonSettings from "./per-lang/makePythonSettings";
 import makeRSettings from "./per-lang/makeRSettings";
+import makeRubySettings from "./per-lang/makeRubySettings";
 import makeRustSettings from "./per-lang/makeRustSettings";
 import makeScalaSettings from "./per-lang/makeScalaSettings.js";
+import makeRacketSettings from "./per-lang/makeRacketSettings.js";
 import makeShellSettings from "./per-lang/makeShellSettings";
 import makeBatchSettings from "./per-lang/makeBatchSettings";
 import makeTsSettings from "./per-lang/makeTsSettings";
@@ -118,8 +123,14 @@ export class SettingsTab extends PluginSettingTab {
 		// ========== TypeScript ==========
 		makeTsSettings(this, this.makeContainerFor("ts"));
 
+		// ========== Lean ==========
+		makeLeanSettings(this, this.makeContainerFor("lean"));
+
 		// ========== Lua ==========
 		makeLuaSettings(this, this.makeContainerFor("lua"));
+
+		// ========== Dart ==========
+		makeDartSettings(this, this.makeContainerFor("dart"));
 
 		// ========== CSharp ==========
 		makeCsSettings(this, this.makeContainerFor("cs"));
@@ -127,18 +138,14 @@ export class SettingsTab extends PluginSettingTab {
 		// ========== Java ==========
 		makeJavaSettings(this, this.makeContainerFor("java"));
 
-
 		// ========== Python ==========
 		makePythonSettings(this, this.makeContainerFor("python"));
-
 
 		// ========== Golang =========
 		makeGoSettings(this, this.makeContainerFor("go"));
 
-
 		// ========== Rust ===========
 		makeRustSettings(this, this.makeContainerFor("rust"));
-
 
 		// ========== C++ ===========
 		makeCppSettings(this, this.makeContainerFor("cpp"));
@@ -148,26 +155,20 @@ export class SettingsTab extends PluginSettingTab {
 
 		// ========== Batch ==========
 		makeBatchSettings(this, this.makeContainerFor("batch"));
-
 		// ========== Shell ==========
 		makeShellSettings(this, this.makeContainerFor("shell"));
-
 
 		// ========== Powershell ==========
 		makePowershellSettings(this, this.makeContainerFor("powershell"));
 
-
 		// ========== Prolog ==========
 		makePrologSettings(this, this.makeContainerFor("prolog"));
-
 
 		// ========== Groovy ==========
 		makeGroovySettings(this, this.makeContainerFor("groovy"));
 
-
 		// ========== R ==========
 		makeRSettings(this, this.makeContainerFor("r"));
-
 
 		// ========== Kotlin ==========
 		makeKotlinSettings(this, this.makeContainerFor("kotlin"));
@@ -175,13 +176,21 @@ export class SettingsTab extends PluginSettingTab {
 		// ========== Mathematica ==========
 		makeMathematicaSettings(this, this.makeContainerFor("mathematica"));
 
-
 		// ========== Haskell ===========
 		makeHaskellSettings(this, this.makeContainerFor("haskell"));
 
 		// ========== Scala ===========
 		makeScalaSettings(this, this.makeContainerFor("scala"));
-		
+
+		// ========== Racket ===========
+		makeRacketSettings(this, this.makeContainerFor("racket"));
+
+		// ========== FSharp ===========
+		makeFSharpSettings(this, this.makeContainerFor("fsharp"));
+
+		// ========== Ruby ============
+		makeRubySettings(this, this.makeContainerFor("ruby"));
+
 		this.focusContainer(this.plugin.settings.lastOpenLanguageTab || canonicalLanguages[0]);
 	}